fix: convert Layout footer to Jahia component for i18n support
The footer in Layout.tsx was rendered as a plain React component, which meant
it was not wrapped by the I18nextProvider injected by the JavaScript Modules
Engine. This caused translations to fail when components from multiple modules
were present on the same page.

The footer has been extracted to a new luxetemplate:footer Jahia component,
ensuring it receives the proper i18n context and translations work correctly
across all modules.

Additionally, updated all server components to use useTranslation() instead of
the direct t() import from i18next for consistency.
