+++ packages/template-set/src/components/JcrQuery/inline.server.tsx
@@ -5,11 +5,11 @@ import {
 	server,
 } from "@jahia/javascript-modules-library";
 import type { JCRNodeWrapper } from "org.jahia.services.content";
-import { t } from "i18next";
 import { buildQuery } from "./utils";
 import type { JcrQueryProps } from "./types";
 import alert from "~/templates/css/alert.module.css";
 import { HeadingSection } from "design-system";
+import { useTranslation } from "react-i18next";
 
 jahiaComponent(
 	{
@@ -33,6 +33,7 @@ jahiaComponent(
 		}: JcrQueryProps,
 		{ currentNode, renderContext },
 	) => {
+		const { t } = useTranslation();
 		const { jcrQuery, warn } = buildQuery({
 			luxeQuery: {
 				"jcr:title": title,
