+++ packages/template-set/src/templates/Footer/definition.cnd
@@ -1,0 +1,1 @@
+[luxetemplate:footer] > jnt:content
