+++ packages/template-set/src/templates/Layout.module.css
@@ -20,33 +20,6 @@ footer {
 	}
 }
 
-.list {
-	composes: listUnstyled from "~/templates/css/global.module.css";
-}
-
-.capitalize {
-	composes: capitalize from "~/templates/css/global.module.css";
-}
-
-.fullTextCapitalize {
-	display: inline-block;
-	text-transform: capitalize;
-}
-
-.disclaimer {
-	font-size: 12px;
-	opacity: 0.6;
-	padding-bottom: 1rem;
-
-	a {
-		font-size: inherit;
-	}
-}
-
-.copyright {
-	text-align: right;
-}
-
 .skipLink {
 	position: absolute;
 	clip: rect(0 0 0 0);
